Create chrony-online service
NetworkManager has a dispatcher which may cause the sources to be marked
as disabled. If you don't have any interfaces actually managed with
NetworkManager, we need to have a way to ensure that the sources get
marked enabled on boot. This change adds a chrony-online service that
waits for the network-online.target in order to be able to make chronyd
source online.

Change-Id: Ib1d91852d281c63b2458fc3c0199a5d5fed17056